pyproject.toml
[tool.poetry]
name = "hw-backend-5"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
attrs = "^23.1.0"
jwt = "^1.3.1"
pydantic = {extras = ["email"], version = "^2.0.3"}


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
